move test funding streams to the test files
