+++ zebra-rpc/qa/rpc-tests/test_framework/config.py
@@ -42,46 +42,3 @@ class ZebraConfig:
             self.extra_args.activation_heights
 
         return config_file
-
-def test_pre_nu6_funding_streams() : return {
-    'recipients': [
-        {
-            'receiver': 'ECC',
-            'numerator': 7,
-            'addresses': ['t26ovBdKAJLtrvBsE2QGF4nqBkEuptuPFZz']
-        },
-        {
-            'receiver': 'ZcashFoundation',
-            'numerator': 5,
-            'addresses': ['t27eWDgjFYJGVXmzrXeVjnb5J3uXDM9xH9v']
-        },
-        {
-            'receiver': 'MajorGrants',
-            'numerator': 8,
-            'addresses': ['t2Gvxv2uNM7hbbACjNox4H6DjByoKZ2Fa3P']
-        },
-    ],
-    'height_range': {
-        'start': 290,
-        'end': 291
-    }
-}
-
-def test_post_nu6_funding_streams() : return {
-    'recipients': [
-        {
-            'receiver': 'MajorGrants',
-            'numerator': 8,
-            'addresses': ['t2Gvxv2uNM7hbbACjNox4H6DjByoKZ2Fa3P']
-        },
-        {
-            'receiver': 'Deferred',
-            'numerator': 12
-            # No addresses field is valid for Deferred
-        }
-    ],
-    'height_range': {
-        'start': 291,
-        'end': 293
-    }
-}
